refactor: hooks 폴더 지정 및 수정
