+++ src/features/TanstackTest/model/hooks/useTanstackQuery.ts
@@ -1,5 +1,5 @@
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
-import { fetchTanstack } from '../api/fetchTanstack';
+import { fetchTanstack } from '@/features/TanstackTest/api/fetchTanstack';
 
 export function useTanstackQuery(): UseQueryResult<any, Error> {
   return useQuery({
